move player into Engine class
